feat: Merge mqtt handler and letsmesh handlers
