+++ repeater/data_acquisition/__init__.py
@@ -1,6 +1,7 @@
-from .mqtt_handler import MQTTHandler
+#from .mqtt_handler import MQTTHandler
 from .rrdtool_handler import RRDToolHandler
 from .sqlite_handler import SQLiteHandler
 from .storage_collector import StorageCollector
 
-__all__ = ["SQLiteHandler", "RRDToolHandler", "MQTTHandler", "StorageCollector"]
+#__all__ = ["SQLiteHandler", "RRDToolHandler", "MQTTHandler", "StorageCollector"]
+__all__ = ["SQLiteHandler", "RRDToolHandler", "StorageCollector"]
